+++ lib/rorvswild/installer.rb
@@ -22,10 +22,12 @@ development:
   # widget: top-left, top-right, bottom-left, bottom-right or hidden
 
   # Open files in your text editor by clicking from the local widget.
-  # VSCode: vscode://file${path}:${line}
-  # Sublime: subl://${path}:${line}
-  # It should be set with an env variable when developers are not using the same editor.
-  editor_url: <%= ENV.fetch("RORVSWILD_EDITOR_URL", "vscode://file${path}:${line}") %>
+  # Leave commented to auto-detect from RAILS_EDITOR or EDITOR env vars (Rails 8.1+).
+  # Or set explicitly:
+  # editor_url: <%= ENV.fetch("RORVSWILD_EDITOR_URL", "vscode://file${path}:${line}") %>
+  # for VSCode: "vscode://file${path}:${line}"
+  # for Sublime: "subl://${path}:${line}"
+  
 
 production:
   api_key: #{api_key}